// asb-good-foods-user/src/app/common/models/order.model.ts
export class Order {
    orderDocId?: string;
    orderItemDocId?: string;
    fullName: string;
    addressLine2: string;
    mobile: number;
    pinCode: number;
    addressLine1: string;
    city: string;
    state: string;
    sameAddress: boolean;
    defaultAddress: boolean;
    sfullName: string;
    saddressLine2: string;
    smobile: number;
    spinCode: number;
    saddressLine1: string;
    scity: string;
    sstate: string;
    productName?: string;
    orderItems?:string;
    description?: string;
    productPrice?: number;
    imageUrl?: string;;
    quantity? : number;
    standard?: boolean;
    tracked?: boolean;
    userId?: string;
    date?: any;
    orderStatus?: string;
    orderId? : number;
    email?: string;
    gst?: number;
    total: number;
    pdf: string;
    orderStatusName?: string;
    quickmode: boolean;
    normalmode: boolean;
    pickupmode:boolean;
    productWeight: number;
    deliveredDate: any;
    orderReqDate:any;
    estimatedDeliveryDate: any;
}